+++ docs/docs/contribution.md
@@ -10,18 +10,20 @@ cd libs/infinity_emb
 poetry install --extras all --with test
 ```
 
-To ensure your contributions pass the Continuous Integration (CI) checks:
+To ensure your contributions pass the Continuous Integration (CI) checks. The `libs/infinity_emb/Makefile` is a useful entrypoint for this.
 ```bash
 cd libs/infinity_emb
 make format
 make lint
 poetry run pytest ./tests
 ```
-As an alternative, you can also use the following command:
+
+As an alternative, you can also use the following command, which bundles a range of the above.
 ```bash
 cd libs/infinity_emb
 make precommit
 ```
 
 ## CLA
+Infinity is developed as open source project. 
 All contributions must be made in a way to be compatible with the MIT License of this repo.
